+++ Sources/ParseSwift/Objects/ParseInstallation+async.swift
@@ -1,0 +1,153 @@
+//
+//  ParseInstallation+async.swift
+//  ParseInstallation+async
+//
+//  Created by Corey Baker on 8/6/21.
+//  Copyright © 2021 Parse Community. All rights reserved.
+//
+
+#if swift(>=5.5) && canImport(_Concurrency)
+import Foundation
+
+@available(macOS 12.0, iOS 15.0, macCatalyst 15.0, watchOS 9.0, tvOS 15.0, *)
+public extension ParseInstallation {
+
+    // MARK: Fetchable - Async/Await
+    /**
+     Fetches the `ParseInstallation` *aynchronously* with the current data from the server
+     and sets an error if one occurs. Publishes when complete.
+     - parameter includeKeys: The name(s) of the key(s) to include that are
+     `ParseObject`s. Use `["*"]` to include all keys. This is similar to `include` and
+     `includeAll` for `Query`.
+     - parameter options: A set of header options sent to the server. Defaults to an empty set.
+     - returns: Returns saved `ParseInstallation`.
+     - throws: `ParseError`.
+     - important: If an object fetched has the same objectId as current, it will automatically update the current.
+    */
+    func fetch(includeKeys: [String]? = nil,
+               options: API.Options = []) async throws -> Self {
+        try await withCheckedThrowingContinuation { continuation in
+            self.fetch(includeKeys: includeKeys,
+                       options: options,
+                       completion: continuation.resume)
+        }
+    }
+
+    // MARK: Savable - Async/Await
+    /**
+     Saves the `ParseInstallation` *asynchronously*.
+     - parameter isIgnoreCustomObjectIdConfig: Ignore checking for `objectId`
+     when `ParseConfiguration.allowCustomObjectId = true` to allow for mixed
+     `objectId` environments. Defaults to false.
+     - parameter options: A set of header options sent to the server. Defaults to an empty set.
+     - returns: Returns saved `ParseInstallation`.
+     - throws: `ParseError`.
+     - important: If an object saved has the same objectId as current, it will automatically update the current.
+    */
+    func save(isIgnoreCustomObjectIdConfig: Bool = false,
+              options: API.Options = []) async throws -> Self {
+        try await withCheckedThrowingContinuation { continuation in
+            self.save(isIgnoreCustomObjectIdConfig: isIgnoreCustomObjectIdConfig,
+                      options: options,
+                      completion: continuation.resume)
+        }
+    }
+
+    // MARK: Deletable - Async/Await
+    /**
+     Deletes the `ParseInstallation` *asynchronously*.
+
+     - parameter options: A set of header options sent to the server. Defaults to an empty set.
+     - returns: Returns saved `ParseInstallation`.
+     - throws: `ParseError`.
+     - important: If an object deleted has the same objectId as current, it will automatically update the current.
+    */
+    func delete(options: API.Options = []) async throws {
+        _ = try await withCheckedThrowingContinuation { continuation in
+            self.delete(options: options, completion: continuation.resume)
+        }
+    }
+}
+
+// MARK: Batch Support - Async/Await
+@available(macOS 12.0, iOS 15.0, macCatalyst 15.0, watchOS 9.0, tvOS 15.0, *)
+public extension Sequence where Element: ParseInstallation {
+    /**
+     Fetches a collection of installations *aynchronously* with the current data from the server and sets
+     an error if one occurs.
+     - parameter includeKeys: The name(s) of the key(s) to include that are
+     `ParseObject`s. Use `["*"]` to include all keys. This is similar to `include` and
+     `includeAll` for `Query`.
+     - parameter options: A set of header options sent to the server. Defaults to an empty set.
+     - returns: Returns saved `ParseInstallation`.
+     - throws: `ParseError`.
+     - important: If an object fetched has the same objectId as current, it will automatically update the current.
+    */
+    func fetchAll(includeKeys: [String]? = nil,
+                  options: API.Options = []) async throws -> [(Result<Self.Element, ParseError>)] {
+        try await withCheckedThrowingContinuation { continuation in
+            self.fetchAll(includeKeys: includeKeys,
+                          options: options,
+                          completion: continuation.resume)
+        }
+    }
+
+    /**
+     Saves a collection of installations *asynchronously*.
+     - parameter batchLimit: The maximum number of objects to send in each batch. If the items to be batched.
+     is greater than the `batchLimit`, the objects will be sent to the server in waves up to the `batchLimit`.
+     Defaults to 50.
+     - parameter transaction: Treat as an all-or-nothing operation. If some operation failure occurs that
+     prevents the transaction from completing, then none of the objects are committed to the Parse Server database.
+     - parameter isIgnoreCustomObjectIdConfig: Ignore checking for `objectId`
+     when `ParseConfiguration.allowCustomObjectId = true` to allow for mixed
+     `objectId` environments. Defaults to false.
+     - parameter options: A set of header options sent to the server. Defaults to an empty set.
+     - returns: - returns: Returns saved `ParseInstallation`.
+     - throws: `ParseError`.
+     - important: If an object saved has the same objectId as current, it will automatically update the current.
+     - warning: If `transaction = true`, then `batchLimit` will be automatically be set to the amount of the
+     objects in the transaction. The developer should ensure their respective Parse Servers can handle the limit or else
+     the transactions can fail.
+    */
+    func saveAll(batchLimit limit: Int? = nil,
+                 transaction: Bool = false,
+                 isIgnoreCustomObjectIdConfig: Bool = false,
+                 options: API.Options = []) async throws -> [(Result<Self.Element, ParseError>)] {
+        try await withCheckedThrowingContinuation { continuation in
+            self.saveAll(batchLimit: limit,
+                         transaction: transaction,
+                         isIgnoreCustomObjectIdConfig: isIgnoreCustomObjectIdConfig,
+                         options: options,
+                         completion: continuation.resume)
+        }
+    }
+
+    /**
+     Deletes a collection of installations *asynchronously*.
+     - parameter batchLimit: The maximum number of objects to send in each batch. If the items to be batched.
+     is greater than the `batchLimit`, the objects will be sent to the server in waves up to the `batchLimit`.
+     Defaults to 50.
+     - parameter transaction: Treat as an all-or-nothing operation. If some operation failure occurs that
+     prevents the transaction from completing, then none of the objects are committed to the Parse Server database.
+     - parameter options: A set of header options sent to the server. Defaults to an empty set.
+     - returns: Returns saved `ParseInstallation`.
+     - throws: `ParseError`.
+     - important: If an object deleted has the same objectId as current, it will automatically update the current.
+     - warning: If `transaction = true`, then `batchLimit` will be automatically be set to the amount of the
+     objects in the transaction. The developer should ensure their respective Parse Servers can handle the limit or else
+     the transactions can fail.
+    */
+    func deleteAll(batchLimit limit: Int? = nil,
+                   transaction: Bool = false,
+                   options: API.Options = []) async throws -> [(Result<Void, ParseError>)] {
+        try await withCheckedThrowingContinuation { continuation in
+            self.deleteAll(batchLimit: limit,
+                           transaction: transaction,
+                           options: options,
+                           completion: continuation.resume)
+        }
+    }
+}
+
+#endif